+++ compiler2/ttcn3/Statement.cc
@@ -6755,7 +6755,7 @@ error:
     str = mputprintf(str, "if (alt_flag == ALT_NO && default_flag == ALT_NO) "
       "TTCN_error(\"Stand-alone %s statement failed in file ", get_stmt_name());
     str = Code::translate_string(str, get_filename());
-    int first_line = get_first_line(), last_line = get_last_line();
+    size_t first_line = get_first_line(), last_line = get_last_line();
     if (first_line < last_line) str = mputprintf(str,
       " between lines %d and %d", first_line, last_line);
     else str = mputprintf(str, ", line %d", first_line);
@@ -13572,7 +13572,7 @@ error:
 	  "TTCN_error(\"None of the branches can be chosen in the alt "
 	  "statement in file ", label_str);
       str = Code::translate_string(str, loc.get_filename());
-      int first_line = loc.get_first_line(), last_line = loc.get_last_line();
+      size_t first_line = loc.get_first_line(), last_line = loc.get_last_line();
       if (first_line < last_line) str = mputprintf(str,
 	" between lines %d and %d", first_line, last_line);
       else str = mputprintf(str, ", line %d", first_line);
@@ -13803,7 +13803,7 @@ error:
     str = mputstr(str, ") TTCN_error(\"None of the branches can be chosen in "
       "the response and exception handling part of call statement in file ");
     str = Code::translate_string(str, loc.get_filename());
-    int first_line = loc.get_first_line(), last_line = loc.get_last_line();
+    size_t first_line = loc.get_first_line(), last_line = loc.get_last_line();
     if (first_line < last_line) str = mputprintf(str,
       " between lines %d and %d", first_line, last_line);
     else str = mputprintf(str, ", line %d", first_line);
